+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `PGS_Construction.createSuperRandomPolygon()` no longer produces holes when `holes` is set to `false`.
 * `PGS_Contour.chordalAxis()` can no longer return polygonal output.
 * `PGS_Voronoi.compoundVoronoi()` now uses the given bounds (previously ignored).
+* `PGS_Processing.generateRandomPoints()` can no longer produce different outputs for the same seed on polygons with holes.
 
 ### Removed
 * `polygonizeLines()` from `PGS_Processing`, in favour of `polygonize(PShape)`.